fix blueprints always disabled
