+++ maps/pirates/coredata.lua
@@ -9,9 +9,6 @@ local Public = {}
 Public.scenario_id_name = 'pirates'
 Public.version_string = '1.6.3' --major.minor.patch versioning, to match factorio mod portal
 
--- Public.blueprint_library_allowed = true
--- Public.blueprint_importing_allowed = true
-
 Public.rocket_silo_death_causes_loss = false
 
 Public.victory_x = 1000
